display: split `DisplayBuffer` and `Display` (#5001)
